ESLint: split config packages
