+++ README.md
@@ -4,18 +4,26 @@ Shared configuration for JavaScript and TypeScript tools.
 
 ## Libraries
 
-| Tool                          | NPM Module                        | Version                                   |
-| ----------------------------- | --------------------------------- | ----------------------------------------- |
-| [ESLint][eslint docs]         | `@viamrobotics/eslint-config`     | [![eslint version][]][eslint npm]         |
-| [Prettier][prettier docs]     | `@viamrobotics/prettier-config`   | [![prettier version][]][prettier npm]     |
-| [TypeScript][typescript docs] | `@viamrobotics/typescript-config` | [![typescript version][]][typescript npm] |
+| Tool                                      | NPM Module                             | Version                                             |
+| ----------------------------------------- | -------------------------------------- | --------------------------------------------------- |
+| [ESLint][eslint docs]                     | `@viamrobotics/eslint-config`          | [![eslint version][]][eslint npm]                   |
+| [ESLint + Svelte][eslint-svelte docs]     | `@viamrobotics/eslint-config-svelte`   | [![eslint-svelte version][]][eslint-svelte npm]     |
+| [Prettier][prettier docs]                 | `@viamrobotics/prettier-config`        | [![prettier version][]][prettier npm]               |
+| [Prettier + Svelte][prettier-svelte docs] | `@viamrobotics/prettier-config-svelte` | [![prettier-svelte version][]][prettier-svelte npm] |
+| [TypeScript][typescript docs]             | `@viamrobotics/typescript-config`      | [![typescript version][]][typescript npm]           |
 
 [eslint docs]: ./packages/eslint-config
 [eslint npm]: https://www.npmjs.com/package/@viamrobotics/eslint-config
 [eslint version]: https://img.shields.io/npm/v/@viamrobotics/eslint-config?style=flat-square
+[eslint-svelte docs]: ./packages/eslint-config-svelte
+[eslint-svelte npm]: https://www.npmjs.com/package/@viamrobotics/eslint-config
+[eslint-svelte version]: https://img.shields.io/npm/v/@viamrobotics/eslint-config?style=flat-square
 [prettier docs]: ./packages/prettier-config
 [prettier npm]: https://www.npmjs.com/package/@viamrobotics/prettier-config
 [prettier version]: https://img.shields.io/npm/v/@viamrobotics/prettier-config?style=flat-square
+[prettier-svelte docs]: ./packages/prettier-config-svelte
+[prettier-svelte npm]: https://www.npmjs.com/package/@viamrobotics/prettier-config-sveltes
+[prettier-svelte version]: https://img.shields.io/npm/v/@viamrobotics/prettier-config-svelte?style=flat-square
 [typescript docs]: ./packages/typescript-config
 [typescript npm]: https://www.npmjs.com/package/@viamrobotics/typescript-config
 [typescript version]: https://img.shields.io/npm/v/@viamrobotics/typescript-config?style=flat-square
