+++ packages/eslint-config-svelte/.npmignore
@@ -1,0 +1,2 @@
+tsconfig.json
+*.tsbuildinfo
