+++ packages/eslint-config/package.json
@@ -4,28 +4,15 @@
     "access": "public",
     "provenance": true
   },
-  "version": "0.7.1",
-  "description": "Common ESLint configuration for Viam projects.",
-  "files": [
-    "**/*",
-    "!tsconfig.json"
-  ],
-  "types": "dist/base.d.cts",
-  "typesVersions": {
-    "*": {
-      "svelte": [
-        "dist/svelte.d.cts"
-      ]
-    }
-  },
+  "version": "0.8.0",
+  "description": "ESLint configuration for projects at Viam.",
+  "type": "commonjs",
+  "main": "./eslint-config.js",
+  "types": "./dist/eslint-config.d.ts",
   "exports": {
     ".": {
-      "types": "./dist/base.d.cts",
-      "default": "./base.cjs"
-    },
-    "./svelte": {
-      "types": "./dist/svelte.d.cts",
-      "default": "./svelte.cjs"
+      "types": "./dist/eslint-config.d.ts",
+      "default": "./eslint-config.js"
     }
   },
   "repository": {
@@ -43,28 +30,17 @@
     "@typescript-eslint/parser": ">=6 <9",
     "eslint": ">=8 <9",
     "eslint-config-prettier": ">=9 <10",
-    "eslint-plugin-jest-dom": ">=5 <6",
     "eslint-plugin-simple-import-sort": ">=12 <13",
-    "eslint-plugin-svelte": ">=2 <3",
-    "eslint-plugin-tailwindcss": ">=3 <4",
-    "eslint-plugin-testing-library": ">=6 <8",
     "eslint-plugin-unicorn": ">=47 <57",
-    "eslint-plugin-vitest": ">=0.5 <0.6"
+    "eslint-plugin-vitest": ">=0.5 <0.6",
+    "typescript": ">=5 <6",
+    "vitest": "*"
   },
   "peerDependenciesMeta": {
-    "eslint-plugin-jest-dom": {
-      "optional": true
-    },
-    "eslint-plugin-svelte": {
-      "optional": true
-    },
-    "eslint-plugin-tailwindcss": {
-      "optional": true
-    },
-    "eslint-plugin-testing-library": {
-      "optional": true
-    },
     "eslint-plugin-vitest": {
+      "optional": true
+    },
+    "vitest": {
       "optional": true
     }
   }
