+++ tsconfig.json
@@ -6,6 +6,7 @@
   },
   "references": [
     { "path": "./packages/eslint-config" },
+    { "path": "./packages/eslint-config-svelte" },
     { "path": "./packages/prettier-config" },
     { "path": "./packages/prettier-config-svelte" }
   ],
